renamed eddl_support to support_eddl
